kbd: reset the keymap before apply customizations
This commit fixes the problem with the `defkeymap' that my
customizations so far have been based on: it does not assign some
keysyms to their intended positions, leaving for example `Meta_colon' or
`Meta_parenleft' unassigned.

This commit adds a shell script to generate the reset keymap from the
output of `dumpkeys', and checks in the generated keymap for easy
installation with Stow. It also removes ad-hoc fixes accumulated in
`eush.map'.