emacs: add configuration for smart-mode-line
If `mode-line-format' configuration does not seem logical, that's
because it's not, and on the default mode line percentage indication
will look strangely duplicated at the beginning of the line. But that's
designed to work around hacks in `smart-mode-line' to ultimately look
rather pretty. (`mode-line-position' is used by `smart-mode-line' as a
placeholder for extra padding from buffer identification, and everything
but percentage indication is moved to `mode-line-front-space'.)

This commit also disables `rich-minority-mode' because it's used
internally by `smart-mode-line', and enabling it explicitly prevents
`smart-mode-line' from processing minor mode lighters properly.