emacs: move time indication to the minibuffer line
This commit moves time indication (previously set up by enabling
`display-time-mode') from the mode line to the minibuffer line, which is
shared between all windows in a frame. This frees up some space in the
mode line and allows to align minor modes list on the right margin in
`smart-mode-line'.