bash: don't adjust SHLVL based on TMUX
If Tmux is exec'd instead of executed as a child process, this
adjustment leads to incorrect SHLVL indication for the subshells.

Removing this check simplifies the config, and in general it seems that
it's better to exec Tmux (or any other terminal multiplexer) if shells
spawned by it are expected to have "SHLVL=1" rather than hack around in
the shell config.