emacs: add more checks in completing-read--w3m-switch-buffer
The advice `completing-read--w3m-switch-buffer' is used to delete the
initial input populated by `w3m-switch-buffer' so that it can be used
nicely with `ivy'. Previously, it only checked `major-mode' which made
it also mistakenly delete initial input from other `completing-read'
calls initiated from a W3M buffer, among which the call in
`read-file-name' used in `w3m-download'.

This commit adds a check that the initial input looks like a W3M buffer
name before proceeding to alter the arguments of the call.