emacs: advice counsel-git-grep to always try counsel-git-grep-projects
If `counsel-git-grep' is passed `t' for `cmd', it will try to resolve
the `default-directory' in `counsel-git-grep-projects' first, but
otherwise it will ask the user for a Git Grep command to run. If it is
passed `nil', it will always use `counsel-git-grep-cmd-default'.

This is not how I prefer it to operate, so this commit advices this
function to first look for a project, but then fall back to
`counsel-git-grep-cmd-default' if `cmd' is `nil'. The behavior is
unchanged if `cmd' is `t'.