emacs: invert meaning of the prefix argument in a project directory
If not in a project directory, prefix argument signals to Counsel to
prompt the user for Git Grep command. When in a project directory,
Counsel was advised to always use project configuration regardless the
value of the prefix argument. This commit makes Counsel invert the
meaning of the prefix argument in this case: if the prefix argument is
set, project configuration is *not* consulted.