emacs: move eval-after-load statements to file init.el
This commit unwraps one layer of nesting in config files by moving
`with-eval-after-load' forms to file `init.el' where possible. This
makes it easier to debug configuration since all defun in these
unwrapped files will be at the top level, but it also may improve
start-up time since the files don't even have to be parsed until
corresponding modes are loaded.

Some files subscribe to several features and hence can't be unwrapped in
this manner. Other files were not (have to be) wrapped in the first
place, so `init.el' contains a mix of both `load' forms and
`with-eval-after-load' forms.