emacs/org: insert links as links in Org capture templates
This changes `org-capture-templates` to insert links as headline links
as opposed to URL properties in property drawers. This makes it easier
to read exported backlogs because links are naturally rendered as links,
and URL properties are not.

Links that are secondary to the content of a headline can be hidden in
the drawer after refiling using `my-org-convert-url-property` introduced
in the previous commit.