emacs/ediff: save window configuration around Ediff sessions
This saves window configuration when entering Ediff, and restores it
when quitting Ediff. Window configurations are saved per frame, so that
configurations from simultaneous Ediff sessions do not interfere with
each other.

This replaces the previous use of `winner-undo` as a `ediff-quit-hook`
which doesn't work when window configuration is changed during a Ediff
sessions (with e.g. `ediff-toggle-split`).