emacs: reimplement my-switch-window as a hydra command
This replaces completion-based implementation of `my-switch-window` with
hydra-based implementation.

The new hydra command supports cycling through active windows in
use-time order and jumping to a particular window by number. Window
numbers are displayed in the hint line and additionally overlayed on top
of the windows themselves, similarly to `swap-buffers` overlays. The
current window is propertized differently from other windows, in the
hint line as well as in overlays.