[emacs/org] Advice org-timestamp-change to file into datetree on no-op
This removes ctrl-c-ctrl-c hook for filing into datetree and replaces it
with an advice around `org-timestamp-change` to offer filing the
headling at point into the datetree whenever the function does not
change the timestamp. This fits the capture-refile process much better
because it activates at the prospective datetree entry, not on the
datetree itself.

OTOH, this also makes `my-org-datetree-file-entries` interactive for
when it is useful.