[emacs/org] Refactor Org extractors
This reduces the concept of an extractor to a function that takes a url
and returns either an Org headline element or nil. Extractor names are
replaced with a function `my-org-url-property-name` that converts any
URL to a property name. Extractor url regexps are now part of the
extractors themselves. All side effects (insertion and exporting via
dynamic variables) are pushed to `my-org-capture-extract-tree`.